+++ src/sqlfluff/core/dialects/dialect_snowflake.py
@@ -66,7 +66,7 @@ snowflake_dialect.sets("unreserved_keywords").update(
         "PIPES",
         "QUERIES",
         "REGIONS",
-        "REMOVE",
+        "SECURE",
         "SEED",
         "TERSE",
         "UNSET",
@@ -172,6 +172,22 @@ class StatementSegment(ansi_StatementSegment):
         Ref("CreateCloneStatementSegment"),
         Ref("ShowStatementSegment"),
         Ref("AlterUserSegment"),
+    )
+
+
+@snowflake_dialect.segment()
+class CreateStatementCommentSegment(BaseSegment):
+    """A comment in a create statement.
+
+    e.g. comment = 'a new view'
+
+    """
+
+    type = "snowflake_comment"
+    match_grammar = Sequence(
+        Ref.keyword("COMMENT"),
+        Ref("EqualsSegment"),
+        Ref("LiteralGrammar"),
     )
 
 
@@ -484,6 +500,7 @@ class CreateStatementSegment(BaseSegment):
             Sequence("EXTERNAL", "TABLE"),
             "VIEW",
             Sequence("MATERIALIZED", "VIEW"),
+            Sequence("SECURE", "VIEW"),
             Sequence("MASKING", "POLICY"),
             "PIPE",
             "FUNCTION",
@@ -501,6 +518,9 @@ class CreateStatementSegment(BaseSegment):
         ),
         Sequence("IF", "NOT", "EXISTS", optional=True),
         Ref("ObjectReferenceSegment"),
+        Ref("CreateStatementCommentSegment", optional=True),
+        Ref.keyword("AS", optional=True),
+        Ref("SelectStatementSegment", optional=True),
     )