+++ src/sqlfluff/core/rules/std/L003.py
@@ -1,6 +1,6 @@
 """Implementation of Rule L003."""
 
-from sqlfluff.core.rules.base import BaseCrawler, LintResult, LintFix
+from sqlfluff.core.rules.base import BaseRule, LintResult, LintFix
 from sqlfluff.core.rules.doc_decorators import (
     document_fix_compatible,
     document_configuration,
@@ -9,7 +9,7 @@ from sqlfluff.core.rules.doc_decorators import (
 
 @document_fix_compatible
 @document_configuration
-class Rule_L003(BaseCrawler):
+class Rule_L003(BaseRule):
     """Indentation not consistent with previous lines.
 
     Note: