+++ src/sqlfluff/core/rules/std/L015.py
@@ -1,11 +1,11 @@
 """Implementation of Rule L015."""
 
-from sqlfluff.core.rules.base import BaseCrawler, LintFix, LintResult
+from sqlfluff.core.rules.base import BaseRule, LintFix, LintResult
 from sqlfluff.core.rules.doc_decorators import document_fix_compatible
 
 
 @document_fix_compatible
-class Rule_L015(BaseCrawler):
+class Rule_L015(BaseRule):
     """DISTINCT used with parentheses.
 
     | **Anti-pattern**