+++ test/api/simple_test.py
@@ -21,6 +21,12 @@ lint_result = [
         "line_pos": 1,
     },
     {
+        "code": "L044",
+        "description": "Query produces an unknown number of result columns.",
+        "line_no": 1,
+        "line_pos": 1,
+    },
+    {
         "code": "L039",
         "description": "Unnecessary whitespace found.",
         "line_no": 1,